+++ modules/generator/processor/servicegraphs/store/store.go
@@ -76,7 +76,7 @@ func (s *store) tryEvictHead() bool {
 // UpsertEdge fetches an Edge from the store and updates it using the given callback. If the Edge
 // doesn't exist yet, it creates a new one with the default TTL.
 // If the Edge is complete after applying the callback, it's completed and removed.
-func (s *store) UpsertEdge(key string, update Callback) (bool, error) {
+func (s *store) UpsertEdge(key string, update Callback) (isNew bool, err error) {
 	s.mtx.Lock()
 	defer s.mtx.Unlock()
 
@@ -93,16 +93,22 @@ func (s *store) UpsertEdge(key string, update Callback) (bool, error) {
 		return false, nil
 	}
 
+	edge := newEdge(key, s.ttl)
+	update(edge)
+
+	if edge.isComplete() {
+		s.onComplete(edge)
+		return true, nil
+	}
+
 	// Check we can add new edges
 	if s.l.Len() >= s.maxItems {
 		// todo: try to evict expired items
 		return false, ErrTooManyItems
 	}
 
-	edge := newEdge(key, s.ttl)
 	ele := s.l.PushBack(edge)
 	s.m[key] = ele
-	update(edge)
 
 	return true, nil
 }